+++ shared_payment_backend/src/main/java/com/practice/shared_payment_backend/models/interfaces/Payment.java
@@ -16,9 +16,9 @@ public interface Payment {
 
     void setDescription(String description);
 
-    int getDate();
+    long getDate();
 
-    void setDate(int date);
+    void setDate(long date);
 
     JsonNode asJson();
 }
